divide wordlist from element_commands
